backend/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
rembg==2.0.59
pillow==12.0.0
gunicorn==23.0.0
numpy==2.3.5
onnxruntime==1.20.1
